# docs/swagger-02.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8020",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/example/{xxl}": {
      "get": {
        "tags": [
          "example-controller"
        ],
        "operationId": "hello",
        "parameters": [
          {
            "name": "xxl",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Successfully checked if chargepoint exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExistsChargepointDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "errorId": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ExistsChargepointDto": {
        "type": "object",
        "properties": {
          "exists": {
            "type": "string"
          }
        }
      }
    }
  }
}